refactor(`CategoryItem`, `PageItem`): move to "src/views/nodes"
