+++ src/views/nodes/categoryItem.ts
